upd
git-svn-id: file:///Users/demichel/SVNREPOS/CODE@4500 fea3b138-1bf5-0310-9da5-cbb00e418fd0

+++ MDsimul++/mdlib/TESTS/timingtest_mp.c++
@@ -42,8 +42,8 @@ using namespace boost::multiprecision::backends;
 using numty=number<mpfr_float_backend<200>>;
 using cmplx=number<mpc_complex_backend<200>>;
 #else
-using numty=long double;
-using cmplx=complex<long double>;
+using numty=double;
+using cmplx=complex<double>;
 #endif
 int main(int argc, char* argv[])
 {
